修正 64-bit architecture 下 code 的一些 warning

+++ e-Hentai/Common/UIAlertView+Hentai.h
@@ -8,7 +8,7 @@
 
 #import <UIKit/UIKit.h>
 
-typedef void (^HentaiClickBlock)(int clickIndex);
+typedef void (^HentaiClickBlock)(NSInteger clickIndex);
 typedef void (^HentaiCancelBlock)(void);
 
 @interface UIAlertView (Hentai) <UIAlertViewDelegate>
